Split struct default semantic validation
